Add class documentation
Principally I agree on the need for defensive copying, but in this case
I would rather document the risk for external mutation, since the copying
could have an impact on performance. And since the primary way of populating
objects of this type is as a result of de-serializing an HTTP entity stream
the risk seems negligible.