Improve handling of MarcReaderException from record service
MarcReaderException can occur when a record can't be parsed by the
MarcReader. In that case we want to tell the user that it is that
specific type of exception and which record it happened for.

MarcReaderException is returned from record service as HTTP 500 which
mean we have to look at the message.